RemoteObjectProxy complete
Move implemention to abstracts.py
Merge tests for attribs and python-jsonschema-objects;
the same tests are executed against both implementations,
with the same remote mock and a similar object design
{"attribute":"value"}
